Refactored the incremental integration tests to take the compilation database as input
