+++ GraphGenerator/Tests/Integration/CMakeLists.txt
@@ -1,8 +1,6 @@
-set(FILES 
-	IncrementalTests.cpp
-	Tests.cpp
+set(INTEGRATIONS 
+	IncrementalTests
 )
-set(TESTS IntegrationTests)
 set(TESTS_LIBS gtest gmock)
 set(SUBDIRECTORIES )
 
@@ -10,9 +8,10 @@ foreach(VAR ${SUBDIRECTORIES})
   add_subdirectory(${VAR})
 endforeach()
 
-if(FILES)
-	add_executable(${TESTS} ${FILES})
-	target_link_libraries(${TESTS} PRIVATE ${PROJECT_NAME} ${TESTS_LIBS})
-endif()
+foreach(VAR ${INTEGRATIONS})
+	add_executable(${VAR} ${VAR}.cpp)
+	target_link_libraries(${VAR} PRIVATE ${PROJECT_NAME} ${TESTS_LIBS})
+endforeach()
+
